Codility Tests

+++ PermMissingElem.java
@@ -35,7 +35,7 @@ import java.util.*;
 class Solution {
     public int solution(int[] A) {
 
-        int N = A.length + 1;
+        long N = A.length + 1;
         long sum = (N * (N + 1)) / 2;
 
         for(int i = 0; i < A.length; i++)
